requirements.txt
# requirements.txt
Django                         
djangorestframework
djangorestframework-simplejwt  # JWT token implementation for DRF
dj-rest-auth[with_social]    # REST API authentication utils
mysqlclient                  # MySQL database adapter for Python
python-dotenv                # For loading environment variables from .env file
requests
drf-yasg                # For generating Swagger documentation